pikabu-test end commit

+++ app/src/main/java/ru/aphanite/pikabu_test/network/PikabuApi.kt
@@ -3,7 +3,6 @@ package ru.aphanite.pikabu_test.network
 import retrofit2.Call
 import retrofit2.http.GET
 import retrofit2.http.Query
-import ru.aphanite.pikabu_test.data.StoryDto
 
 
 interface PikabuApi {
@@ -12,5 +11,5 @@ interface PikabuApi {
     fun getAllStories(): Call<List<StoryDto>>
 
     @GET("story.php")
-    fun getStoryById(@Query("id") id: Int): Call<StoryDto>
+    fun getStoryById(@Query("id") id: Long): Call<StoryDto>
 }
